### 2020-02-24 #369 @mcguinlu
Presubmission inquiry: medrxivr
Submitting Author: Luke McGuinness (@mcguinlu)
Repository: https://github.com/mcguinlu/medrxivr

---

```
Package: medrxivr
Title: Access MedRxiv Data In R
Version: 0.0.1.900
    person("Luke", "McGuinness",
           email = "luke.mcguinness@bristol.ac.uk",
    person("Lena", "Schmidt",
           role = "aut",
           comment = c(ORCID = "0000-0003-0709-8226")))
Description: medRxiv (https://www.medrxiv.org/) is a free online archive and
    distribution server for complete but unpublished manuscripts (preprints) in
    the medical, clinical, and related health sciences. medrxivr provides
    programmatic access to a snapshot of the preprints contained on medRxiv,
    which is updated daily. Users can search for relevant records using regular
    expressions and Boolean logic, and can easily download the full-text PDFs
    of preprints matching their search criteria.
    rvest,
    xml2
    covr,
    kableExtra
```

- Please indicate which category or categories from our [package fit policies](https://ropensci.github.io/dev_guide/policies.html#package-categories) this package falls under: (Please check an appropriate box below.:

`medrxivr` allows users to programmatically access and manipulate a snapshot of [medRxiv](https://www.medrxiv.org/), a preprint respository for papers in medical, clinical, and related health sciences. The snapshot is automatically updated each morning (webscraping/cleaning script can be found [here](https://github.com/mcguinlu/autosynthesis/blob/master/R/extract_daily.R)).

The primary target of this package is systematic reviewers (i.e. me!), who frequently wish both to use more complicated queries (e.g. regular expresssions/Boolean combinations) when searching medRxiv than the official site currrently allows for, and who also wish to be easily able to download the full text PDFs of records matching their search. `medrxivr` helps with both of these challenges. However, anyone who wishes to find and retrieve relevant medRxiv records in R, for example to explore the distribution of preprints by subject area, will find the package useful.

As far as I am aware, no other package allows users to access medRxiv data in R.

`medrxivr` is not yet ready for submission - there is some additional functionality that I would like to incorporate into the package and I need to fix one or two issues with the webscraping/data cleaning script. However, I did have two questions it would be great to get feedback on:
    - Q1: `medrxivr` provides access to a static snapshot of the medRxiv data, which I also maintain, in addition to functions to search for and download relevant full text records. The reason this approach was taken, rather than providing tools for users to search the site dynamically, is that the [`robots.txt`](https://www.medrxiv.org/robots.txt) on medRxiv forbids the scraping of the `search/` path. In this case, providing functions that will query this path multiple times seemed like a bad idea. However, I want to check that my setup - a package providing access to a snapshot of the data, both of which are maintained by the same person - is an acceptable setup for an rOpenSci package?
    - Q2: In terms of intergration with other rOpenSci packages, it seems like `medrxivr` could be a useful addition to [`fulltext`](https://github.com/ropensci/fulltext). Is it worth contacting the maintainers of this package about intergration before or after submitting `medrxivr` to the peer review process?

[Tagging my co-author @L-ENA for reference]
