### 2020-02-25 #369 @annakrystalli
Hello @mcguinlu and many thanks for your presubmission enquiry!

In general, we would consider the package in scope. However, there are  couple of questions we'd like a bit more info on:

1. Where are the data snapshot stored?
2. Are the medRxiv website maintainers aware of the package's use of their data? Especially given the data snapshot uses web scraping and seems undocumented. We would need confirmation that they are happy for your package to use their data as the package does to accept it.

Here are also some initial comments too:
1. 2. Given the package accesses a static snapshot of the data, the README is a bit misleading since it doesn't mention that at all.
2. We recommend you make contact with [`fulltext`](https://github.com/ropensci/fulltext) authors to assess potential compatibility

I look forward to some clarification in response to our queries.
