### 2020-02-25 e1258a4
Update DESCRIPTION and README
