### 2020-02-25 5dbeeaa
Fix vignette build issue on TRAVIS
