### 2020-03-02 7c51f10
Updated tests to improve test coverage
