### 2020-03-02 c1d54ce
Move mx_info() to top to provide immediate feedback to users
