### 2020-03-02 31473a1
Range of updates Move mx_info() to top Fix code behind depulpication option which was removing the version numbers from outputted records Rewrite "*" code to prevent it from creating a warning
