### 2020-03-02 f442da3
Update mx_search() documentation
