### 2020-03-02 a83653b
Added global binding for link_group variable
