+++ README.Rmd
@@ -30,7 +30,6 @@ knitr::opts_chunk$set(
 
 You can install the development version of this package using:
 
-
 ``` {r, eval = FALSE}
 devtools::install_github("mcguinlu/medrxivr")
 library(medrxivr)
@@ -38,7 +37,7 @@ library(medrxivr)
 
 ## Example
 
-To get the entire _medRxiv_ dataset to play around with, use the following command:
+To get the entire most recent _medRxiv_ snapshot to play around with, use the following command:
 
 ``` {r, eval=FALSE}
 
@@ -115,4 +114,8 @@ mx_download(mx_results,     # Object returned by mx_search
 Please note that the `medrxivr` project is released with a
 [Contributor Code of Conduct](CODE_OF_CONDUCT.md).
 By contributing to this project, you agree to abide by its terms.
+
+## Disclaimer
+
+This package and the data it accesses/returns are provided "as is", with no guarantee of accuracy. Please be sure to check the accuracy of the data yourself (and do let me know if you find an issue so I can fix it for everyone!)
 
