### 2020-03-02 345fb13
Remove redundant code
