### 2020-03-02 f12d966
Add tests to check that adding new topics reduces number of hits rather than increases it
