### 2020-03-02 #369 @mcguinlu
Hi @annakrystalli, thanks for getting back to me, and for your inital comments!

Great to hear that the pacakge is provisionally in scope - I've provided responses to the four queries you raised below:

1. The snapshot is currently stored as a CSV file in the same repository as the webscraping script (see [here](https://github.com/mcguinlu/autosynthesis/blob/master/data/medRxiv_abstract_list.csv)). I appreciate this is a basic solution, but this is one of the areas I was hoping to solict reviewer feedback on, as I have never had to host a dataset before and am unsure what the best solution is.

2. As far as I know, the medRxiv maintainers are not currently aware of this package, though I have now sent an email querying whether they are happy for me to provide access to the data in this way. I really hope they do, as the original arXiv repository takes a [strong view](https://arxiv.org/help/bulk_data) that making the data available programmatically is a cornerstone of open access publishing. I will let you know when I hear back.

3. I completely agree with this comment - the original README introduction was written before I had hammered out the functionality. I've gone through the package (README,  documentation, and function messages) to ensure that it is clear `medrixr` provides access to a static snapshot rather than dynamic access to the repository.

4. I have opened an [issue](https://github.com/ropensci/fulltext/issues/213) on the `fulltext` repository to begin a discussion about the possibility of incorporating `medrixr` as an additional data source.

Thanks again for your inital feedback, and do let me know if I can provide any further clarifications!
