### 2020-03-02 #369 @annakrystalli
Thanks for the clarifications @mcguinlu !

> As far as I know, the medRxiv maintainers are not currently aware of this package, though I have now sent an email querying whether they are happy for me to provide access to the data in this way. I really hope they do, as the original arXiv repository takes a strong view that making the data available programmatically is a cornerstone of open access publishing. I will let you know when I hear back.

Great and yes, I imagine they will be fine with it. They even be able to provide better access to the data rather than crawling their site.

> The snapshot is currently stored as a CSV file in the same repository as the webscraping script (see here). I appreciate this is a basic solution, but this is one of the areas I was hoping to solict reviewer feedback on, as I have never had to host a dataset before and am unsure what the best solution is.

Depending on the size of the files, I wonder if using [pins](http://pins.rstudio.com/articles/use-cases.html) might be a good solution for this problem. You could also consider getting feedback by posting on [rOpenSci Discuss](https://discuss.ropensci.org/). It's quite an interesting topic.

> I completely agree with this comment - the original README introduction was written before I had hammered out the functionality. I've gone through the package (README, documentation, and function messages) to ensure that it is clear medrixr provides access to a static snapshot rather than dynamic access to the repository.
👍

> I have opened an issue on the fulltext repository to begin a discussion about the possibility of incorporating medrixr as an additional data source.

I saw that. Nice work! Will look out for any feedback.

We're pretty much satisfied with response to our comments at this stage, so let's see what the feedback is from your enquiries.
