+++ README.Rmd
@@ -109,6 +109,11 @@ mx_download(mx_results,     # Object returned by mx_search
 
 ```
 
+## Planned updates
+
+ * At present, only the title and author fields are searched. I would like to also make the author/topic fields searchable.
+ * At present, only the first author for each record has been extracted. This will be corrected in future updates.
+
 ## Code of conduct
 
 Please note that the `medrxivr` project is released with a
