### 2020-03-03 4bd6da0
Change NOT default to "" rather than NULL Needed for it to work with the Shiny app
