### 2020-03-04 8173ab3
Fix issue with magrittr dependency
