+++ NAMESPACE
@@ -4,7 +4,7 @@ export(mx_crosscheck)
 export(mx_download)
 export(mx_info)
 export(mx_search)
-importFrom(magrittr,"%>%")
+importFrom(dplyr,"%>%")
 importFrom(methods,is)
 importFrom(stats,runif)
 importFrom(utils,download.file)
