+++ DESCRIPTION
@@ -26,7 +26,6 @@ Imports:
     rvest,
     methods,
     dplyr,
-    magrittr,
     stringr,
     xml2
 Suggests: 
