### 2020-03-04 d3b21e9
Add file encoding so it works with app
