+++ README.Rmd
@@ -14,7 +14,8 @@ knitr::opts_chunk$set(
 )
 ```
 
-# medrxivr
+# medrxivr <img src="man/figures/hex-medrxivr.png" align="right" width="20%" height="20%" />
+
 
 <!-- badges: start -->
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
