### 2020-03-14 #369 @annakrystalli
Hey @mcguinlu,

Was just wondering whether there were any updates on your contact with the medRxiv maintainers? I'm going to then close the issue after that.

Also, regarding your data caching challenge,  you might like to have a look at how the developers of [`ramlegacy`](https://github.com/ropensci/ramlegacy) approached it.
